Default xsoar

+++ demisto_sdk/commands/content_graph/objects/pack.py
@@ -541,10 +541,16 @@ class Pack(BaseContent, PackMetadata, content_type=ContentType.PACK):  # type: i
         documentation_path = CONTENT_PATH / "Documentation"
         documentation_output = destination_path / "Documentation"
         documentation_output.mkdir(exist_ok=True, parents=True)
-        shutil.copy(
-            documentation_path / f"doc-howto-{marketplace.value}.json",
-            documentation_output / "doc-howto.json",
-        )
+        if (documentation_path / f"doc-howto-{marketplace.value}.json").exists():
+            shutil.copy(
+                documentation_path / f"doc-howto-{marketplace.value}.json",
+                documentation_output / "doc-howto.json",
+            )
+        else:
+            shutil.copy(
+                documentation_path / "doc-howto-xsoar.json",
+                documentation_output / "doc-howto.json",
+            )
         if (documentation_path / "doc-CommonServer.json").exists():
             shutil.copy(
                 documentation_path / "doc-CommonServer.json",
